example/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-extensions==3.2.3
python-decouple==3.8
Pillow==10.0.1
django-haystack==3.2.1
whoosh==2.7.4
requests==2.31.0
django-auth-ldap==4.6.0
django-axes==6.1.1
python-dotenv==1.0.0
